ksafe desktop file setup
