+++ foliary/src/jvmMain/kotlin/dev/appoutlet/foliary/core/database/DatabaseModule.jvm.kt
@@ -4,7 +4,7 @@ package dev.appoutlet.foliary.core.database
 
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import ca.gosyer.appdirs.AppDirs
+import dev.appoutlet.foliary.core.appdirs.Directories
 import org.koin.core.annotation.Module
 import org.koin.core.annotation.Single
 import java.io.File
@@ -13,27 +13,10 @@ import java.io.File
 actual object PlatformDatabaseModule {
     @Single
     fun provideDatabaseBuilder(): RoomDatabase.Builder<FoliaryDatabase> {
-        val databaseFolder = getUserDataFolder()
-        createFolderIfItDoesntExists(databaseFolder)
-        val dbFile = File(databaseFolder, "Foliary.db")
+        val databaseFolder = Directories.userDirectory
+        val databaseFile = File(databaseFolder, "Foliary.db")
         return Room.databaseBuilder<FoliaryDatabase>(
-            name = dbFile.absolutePath,
+            name = databaseFile.absolutePath,
         )
-    }
-
-    private fun getUserDataFolder(): String {
-        val appDirs = AppDirs {
-            appName = "Foliary"
-            appAuthor = "AppOutlet"
-        }
-
-        return appDirs.getUserDataDir(roaming = true)
-    }
-
-    private fun createFolderIfItDoesntExists(databaseFolder: String) {
-        val file = File(databaseFolder)
-        if (!file.exists()) {
-            file.mkdirs()
-        }
     }
 }
