+++ foliary/src/jvmMain/kotlin/dev/appoutlet/foliary/core/encryption/EncryptionModule.jvm.kt
@@ -1,5 +1,6 @@
 package dev.appoutlet.foliary.core.encryption
 
+import dev.appoutlet.foliary.core.appdirs.Directories
 import eu.anifantakis.lib.ksafe.KSafe
 import org.koin.core.annotation.Configuration
 import org.koin.core.annotation.Module
@@ -13,5 +14,11 @@ actual object EncryptionModule {
 }
 
 private object KSafeWrapper{
-    val ksafe = KSafe()
+    val ksafe by lazy {
+        KSafe(
+            fileName = "foliary",
+            lazyLoad = true,
+            baseDir = Directories.userDirectory
+        )
+    }
 }
